set the type aliases back to normal

+++ src/Util/Types.h
@@ -2,8 +2,8 @@
 
 #include <cstdint>
 
-using u8 = int;
-using u16 = int;
+using u8 = uint8_t;
+using u16 = uint16_t;
 
-using i8 = int;
-using i16 = int;
+using i8 = int8_t;
+using i16 = int16_t;
